Skip confirmation on invite and confirm on accept
The confirmable model sets the confirmed_at value when you call
skip_confirmation! and this means that a user which has yet to
accept the inviate and considered confirmed.

Instead, I've use hacky workaround so that a new record being invited will
respond with false to confirmation_required? and thereby not sending
a confirmation email on creation.