Updating readme to add accept_user_invitation_url correctly
The readme in itself was a great help getting started for bypassing and sending my own emails to invite users.
However, it was lacking the portion to create my own accept_user_invitation_url correctly. 

invitation_token and the raw_invitation_token variables are a litle vague.
raw_invitation_token is actually the temporarily encrypted token which is used for the url and decrypted. 

it confused me anyway - I hope this clears things up for others ;-) 
thanks