Update README.rdoc
I attempted to send a customized email in a resque job, I found out the when I rebuilt my user object, that .raw_invitation_token wasn't available.  This issue is likely to come up again with Rails 4.2 activejob.  

The steps to recreate the situation is:

1. Create user  user = User.new
2. user.skip_invitation = true
3. user.invite!(so_and_so)
4. user.save
5. Resque.enqueue(SendInvite, user.id.to_s)

the SendInvite will then send the invite and won't have access to @raw_invitation_token.

Thank you!

-daniel