Yield `resource` before validity/acceptance checks
In certain circumstances, I need to allow the user to do more than
simply set a password when accepting an invite. A simple example is to
allow the user to set their name etc. If this is required by the model
validations, it currently breaks.

By checking if for validation errors (`invitation_accepted`) after the
yield, then we can use the block to complete the setup of the resource. This
prevents the need to rewrite the controller in our own code, and we can
do a lot more with the block.