Switch to GitHub Actions
This commit adds CI using GitHub Actions to replace the now non-functional Travis CI.

In addition to configuring CI this commit:

Adds Rubies - 2.7, 3.0, 3.1, jruby-9.3 - to the CI matrix
Replaces the deprecated (and eventually removed) use of update_attributes with update for ActiveRecord
Adds gemfiles for Rails 6.0, 6.1, 7.0
Removes the *.lock files in the gemfiles subdirectory, and adds them to .gitignore
Changes the Rails-master to Rails-main in the gemfiles, in accordance with current Rails branch naming
Replaces the use of classes as arguments to Devise.mailer = with class names, as is expected in Rails 7.
Adds migration support for Rails 6+ to the ActiveRecord configuration

Certain combinations are excluded from the CI matrix because of compatibility issues.  This includes the use of mongoid with Rails 7 and/or Ruby 3.1, because neither is officially supported yet.

ruby-head was removed from the support matrix because we're seeing failures inside Thor that appear to be unrelated to this gem.  Once the Thor issue is identified and fixed, ruby-head can be readded to the matrix